break out into component files and update to only use stock MUI components
